написаны все контроллеры в основном сервисе, кроме Public: Подборки событий Admin: Подборки событий

+++ main-service/src/main/java/explore_with_me/services/adminServices/AdminCategoryService.java
@@ -1,27 +1,29 @@
-package explore_with_me.services;
+package explore_with_me.services.adminServices;
 
+import explore_with_me.exeption.ConflictDataException;
 import explore_with_me.exeption.DataNotFound;
 import explore_with_me.models.category.Category;
 import explore_with_me.models.category.CategoryDto;
 import explore_with_me.models.category.CategoryMapper;
 import explore_with_me.models.category.NewCategoryDto;
+import explore_with_me.models.event.Event;
 import explore_with_me.repositories.CategoryRepository;
+import explore_with_me.services.publicServices.PublicEventService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
-public class CategoryService {
+public class AdminCategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final PublicEventService publicEventService;
 
     @Autowired
-    public CategoryService(CategoryRepository categoryRepository) {
+    public AdminCategoryService(CategoryRepository categoryRepository, PublicEventService publicEventService) {
         this.categoryRepository = categoryRepository;
+        this.publicEventService = publicEventService;
     }
 
     public CategoryDto addNewCategory(NewCategoryDto newCategoryDto) {
@@ -37,24 +39,12 @@ public class CategoryService {
 
     public void deleteCategory(Long catId) {
         findCategory(catId);
-        //TODO написать проверку, что в категории нет ни одного события, иначе исключение.
+        List<Event> events = publicEventService.getEventsByCategoryId(catId);
+        if(events.size() < 1) {
+            throw new ConflictDataException(String.format("Категория, в которой есть события не может быть удалена. " +
+                    "В категории с id = %d есть собыия со следующими id: %s", catId, events));
+        }
         categoryRepository.deleteById(catId);
-    }
-
-    public List<CategoryDto> getCategories(Integer from, Integer size) {
-        int page = from / size;
-        PageRequest pageRequest = PageRequest.of(page, size, Sort.by("name"));
-        return categoryRepository.findAll(pageRequest).stream()
-                .map(CategoryMapper::toCategoryDto).collect(Collectors.toList());
-    }
-
-    public CategoryDto getCategoryById(Long catId) {
-        return CategoryMapper.toCategoryDto(findCategory(catId));
-    }
-
-    public Category findOrCreateCategory(CategoryDto categoryDto) {
-        return categoryRepository.findById(categoryDto.getId()).
-                orElse(categoryRepository.save(CategoryMapper.toCategory(categoryDto)));
     }
 
     private Category findCategory(Long id) {
